tools/requirements.txt
google-genai>=0.3.0
python-pptx>=1.0.0
python-docx>=1.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0
